+++ main.cpp
@@ -147,7 +147,7 @@ int main(int argc, char *argv[])
 		}
 	}
 	std::string target("");
-	int tn_len = strlen(argv[1]);
+	size_t tn_len = strlen(argv[1]);
 	if(argv[1][tn_len - 4] != '.') {
 		target = string_format("%s.lua", argv[1]);
 	} else {
